fix: cascade delete cliente + testdata

+++ src/main/java/com/example/demo/service/ClienteService.java
@@ -55,7 +55,9 @@ public class ClienteService {
     }
 
     public void deleteCliente(Integer id) {
-        pedidoService.deletePedido(pedidoRepository.findByClienteId(id).getId());
+        if (pedidoRepository.findByClienteId(id) != null) {
+               pedidoService.deletePedido(pedidoRepository.findByClienteId(id).getId());
+        }
         clienteRepository.deleteById(id);
     }
 }
